Add view model for system status report view
